backend/requirements.txt
# FastAPI
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic-settings==2.2.1

# MongoDB
motor==3.6.0
beanie==1.26.0
pymongo==4.9.2

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ML / Computer Vision
ultralytics==8.2.0
onnxruntime==1.24.4
opencv-python==4.13.0.92
numpy>=2.0
Pillow>=11.0

# Utilities
python-dotenv==1.0.1
